gst/vaapi: move VP8 encoder test classes to lib
